Improved: Move SeoConfigUtil.java from Product to Widget component (OFBIZ-12031)
Help to remove dependency of product component from widget component
An afterthought: webapp component seems better for SeoConfigUtil.java than widget component.

Thanks Jacques for review
