+++ applications/product/src/main/java/org/apache/ofbiz/product/category/SeoUrlUtil.java
@@ -19,6 +19,7 @@
 package org.apache.ofbiz.product.category;
 
 import org.apache.ofbiz.base.util.UtilValidate;
+import org.apache.ofbiz.webapp.SeoConfigUtil;
 
 public final class SeoUrlUtil {
